+++ packages/account-core/src/index.ts
@@ -10,6 +10,7 @@ export {
   getEthereumWallets,
 } from "@dynamic-labs/global-wallet-client/features";
 export { createEIP1193Provider } from "@dynamic-labs/global-wallet-client/ethereum";
+export * from "./abis";
 export { isEthereumWallet } from "@dynamic-labs/ethereum";
 export { isZKsyncConnector } from "@dynamic-labs/ethereum-aa-zksync";
 export { SiwsMessage } from "./siws";