+++ packages/account-core/tsup.config.ts
@@ -3,7 +3,13 @@ import path from "path";
 
 export default defineConfig((options) => {
   return {
-    entry: ["src/index.ts", "src/wallet.ts", "src/config.ts", "src/siws.ts"],
+    entry: [
+      "src/index.ts",
+      "src/wallet.ts",
+      "src/config.ts",
+      "src/siws.ts",
+      "src/abis/*",
+    ],
     splitting: true,
     sourcemap: !!options.watch,
     clean: true,