+++ packages/account-react/src/index.ts
@@ -3,3 +3,5 @@ export * from "./hooks";
 export * from "./components";
 export * from "./formatters";
 export * from "./token";
+export { IsBrowser } from "@dynamic-labs/sdk-react-core";
+export { ProviderEnum } from "@dynamic-labs/types";