+++ packages/account-core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sophon-labs/account-core
 
+## 1.1.16
+
+### Patch Changes
+
+- Fixing wrong import
+
+## 1.1.15
+
+### Patch Changes
+
+- Updating wrong import
+
+## 1.1.14
+
+### Patch Changes
+
+- CSS Changes and better hook inputs
+
+## 1.1.13
+
+### Patch Changes
+
+- Exposing more methods
+
 ## 1.1.12
 
 ### Patch Changes